Divide id generators in stores/slices
