+++ src/__tests__/redux/point.reducers.test.ts
@@ -1,6 +1,7 @@
 import {store} from 'stores/index';
 import {EntityStateWithLastID} from 'stores/slices/EntityStateWithLastID';
-import {point_slice,generate_point_id} from 'stores/slices/point';
+import {point_slice} from 'stores/slices/point';
+import {generate_point_id} from 'stores/slices/id_generators';
 import {mPoint} from 'models/mPoint';
 
 const initial_point_state: EntityStateWithLastID<mPoint>={
